First Major Update
The Conversation logic is now very efficient.
A faste answear button speeds up the response time and still provides excelent information quality.
Text is now readable and sorted well.
A colour shema is used to make the text easy to read and in red the current text that is read is highlighted.
Scrolling and resizing issus of the UI are fixed.