Added Wisper Support
In the settings wisper support can be enabled. However currently it listens until the listening time is passed. So in the next commit this needs to be fixed so the recording is stopped and started when the user starts and stops speaking for the wisper stt system to be useful